backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.0.0
tensorflow-cpu>=2.12.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
opencv-python-headless>=4.8.0
pandas>=2.0.0
anthropic>=0.25.0
supabase>=2.0.0
